+++ src/App.js
@@ -1,15 +1,15 @@
 import './App.css';
 import React,{useState,useEffect} from 'react';
-import Login from './Components/Login'
-import Contact from './components/contact/Contact';
+import Login from './components-2/Login'
+import Contact from './components-1/contact/Contact';
 import { HashRouter as Router, Routes, Route } from 'react-router-dom';
-import Appointment from './components/appointment/Appointment';
-import CarListing from './components/hire/CarListing';
-import Details from './components/pages/Details';
-import Error from './components/error-page/Error';
-import Payment from './components/pages/payment/CardPayment';
-import AllPayment from './components/pages/payment/AllPayments';
-import Mpesa from './components/pages/payment/Mpesa';
+import Appointment from './components-1/appointment/Appointment';
+import CarListing from './components-1/hire/CarListing';
+import Details from './components-1/pages/Details';
+import Error from './components-1/error-page/Error';
+import Payment from './components-1/pages/payment/CardPayment';
+import AllPayment from './components-1/pages/payment/AllPayments';
+import Mpesa from './components-1/pages/payment/Mpesa';
 
 import Home from './Home /Home';
 import About from './Home /About';
@@ -22,19 +22,19 @@ import "slick-carousel/slick/slick-theme.css";
 
 
 function App() {
-  const [user, setUser] = useState(null);
+  // Authentication Part
 
-  useEffect(() => {
-    // auto-login
-    fetch("/me").then((r) => {
-      if (r.ok) {
-        r.json().then((user) => setUser(user));
-      }
-    });
-  }, []);
+  // const [user, setUser] = useState(null);
+  // useEffect(() => {
+  //   // auto-login
+  //   fetch("/me").then((r) => {
+  //     if (r.ok) {
+  //       r.json().then((user) => setUser(user));
+  //     }
+  //   });
+  // }, []);
 
-
-  if (user) return <Login onLogin={setUser} />;
+  // if (!user) return <Login onLogin={setUser} />;
 
   return (
     <div className="App">
