+++ src/Home /Blog.js
@@ -3,8 +3,8 @@ import { Container, Row } from "reactstrap";
 import Helmet from "../mycomponents/Helmet/Helmet";
 import CommonSection from "../mycomponents/UI/CommonSection";
 import BlogList from "../mycomponents/UI/BlogList";
-import Navbar from "../components/navbar/Navbar";
-import Footer from "../components/footer/Footer";
+import Navbar from "../components-1/navbar/Navbar";
+import Footer from "../components-1/footer/Footer";
 
 const Blog = () => {
   return (
