+++ src/components-1/pages/Details.js
